+++ flexiapi/app/Http/Controllers/Api/Admin/Account/CardDavCredentialsController.php
@@ -46,10 +46,10 @@ class CardDavCredentialsController extends Controller
         $accountCarddavCredentials->space_carddav_server_id = $cardDavServer->id;
         $accountCarddavCredentials->account_id = $account->id;
         $accountCarddavCredentials->username = $request->get('username');
-        $accountCarddavCredentials->domain = $request->get('domain');
+        $accountCarddavCredentials->realm = $request->get('realm');
         $accountCarddavCredentials->password = bchash(
             $request->get('username'),
-            $request->get('domain'),
+            $request->get('realm'),
             $request->get('password'),
             $request->get('algorithm')
         );
@@ -70,7 +70,7 @@ class CardDavCredentialsController extends Controller
         return [
             'carddav_id' => $cardDavServer->id,
             'username' => $cardDavServer->pivot->username,
-            'domain' => $cardDavServer->pivot->domain,
+            'realm' => $cardDavServer->pivot->realm,
             'algorithm' => $cardDavServer->pivot->algorithm,
             'password' => $cardDavServer->pivot->password,
         ];